move ruport class eval fun to an initializer
